fix Second

+++ domain/domain-object/src/main/java/com/example/domainObject/game/game/Second.kt
@@ -4,4 +4,4 @@ package com.example.domainObject.game.game
 /**
  * 秒数
  */
-data class Second(val value: Int)
+data class Second(val value: Long)
